requirements.txt
# ============================================
# Doctor Assistant - Requirements
# Multi-Agent Medical System with Gradio UI
# ============================================

# -----------------------------
# Core LLM & Agent Framework
# -----------------------------
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# -----------------------------
# Vector Stores & Embeddings
# -----------------------------
chromadb>=0.5.0
faiss-cpu>=1.8.0
# faiss-gpu>=1.8.0  # Uncomment if using GPU

# -----------------------------
# MCP (Model Context Protocol)
# -----------------------------
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0

# -----------------------------
# Database Clients
# -----------------------------
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# -----------------------------
# Configuration & Validation
# -----------------------------
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# -----------------------------
# HTTP & API Clients
# -----------------------------
httpx>=0.27.0
aiohttp>=3.9.0
requests>=2.31.0

# -----------------------------
# Web UI - Gradio
# -----------------------------
gradio>=5.0.0
gradio-client>=1.0.0

# -----------------------------
# CLI Interface
# -----------------------------
typer[all]>=0.12.0
rich>=13.7.0              # Rich console output

# -----------------------------
# Utilities
# -----------------------------
tenacity>=8.2.0           # Retry logic
structlog>=24.1.0         # Structured logging
python-json-logger>=2.0.0

# -----------------------------
# Math & Data Processing
# -----------------------------
numpy>=1.26.0
pandas>=2.2.0

# -----------------------------
# Async Support
# -----------------------------
anyio>=4.3.0
asyncio-throttle>=1.0.0

# -----------------------------
# Testing
# -----------------------------
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-env>=1.1.0
respx>=0.21.0             # Mock httpx requests
faker>=24.0.0             # Generate fake data

# -----------------------------
# Development Tools
# -----------------------------
black>=24.2.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0
ruff>=0.3.0               # Fast Python linter

# -----------------------------
# Documentation
# -----------------------------
mkdocs>=1.5.0
mkdocs-material>=9.5.0

# -----------------------------
# Graph Visualization
# -----------------------------
matplotlib>=3.8.0
pillow>=10.2.0            # Image processing for UI

# -----------------------------
# Optional: Monitoring & Tracing
# -----------------------------
langsmith>=0.1.0        # LangChain tracing
# opentelemetry-api>=1.23.0
# opentelemetry-sdk>=1.23.0


pypdf

langchain_huggingface
sentence-transformers

pip install langchain-huggingface==0.1.0 sentence-transformers